Create updated_prime_checker
we are verifying if the given number is prime or not and for that we are using list comprehension to store the divisors. if that list_divisor is empty then number is prime and if not empty then number is not prime and will show all the divisors in a list called list_divisor.